Fix: Auth 관심사 분리
